NVSHAS-8852 Risk->vuln page, move the accept button to each vul entry
