+++ admin/webapp/websrc/app/routes/vulnerabilities/vulnerability-items/vulnerability-items-table/vulnerability-items-table-action-cell/vulnerability-items-table-action-cell.component.html
@@ -1,11 +1,40 @@
 <ng-container *ngIf="params.node.data">
-  <button
-    (click)="downloadCsv()"
-    aria-label="Download CSV icon button "
-    mat-icon-button>
-    <i class="eos-icons">download</i>
-  </button>
-  <button (click)="showInfo()" aria-label="CVE info button" mat-icon-button>
-    <i class="eos-icons">info</i>
-  </button>
+  <div style="display: flex">
+    <button
+      (click)="downloadCsv()"
+      aria-label="Download CSV icon button "
+      mat-icon-button
+    >
+      <i class="eos-icons">download</i>
+    </button>
+    <button (click)="showInfo()" aria-label="CVE info button" mat-icon-button>
+      <i class="eos-icons">info</i>
+    </button>
+    <div *appDisplayControl="'write_vuls_profile'">
+      <button
+        [disabled]="cfgType === 'ground'"
+        (click)="acceptVulnerability()"
+        aria-label="Accept the Vulnerability button"
+        mat-icon-button
+      >
+        <i
+          *ngIf="!isProcessingVulnerabilityAcceptance"
+          [matTooltip]="
+            cfgType === 'ground'
+              ? 'cveProfile.CRD_ACCEPT'
+              : ('cveProfile.ACCEPT' | translate)
+          "
+          matTooltipPosition="above"
+          [matTooltipClass]="'custom-tooltip'"
+          class="eos-icons"
+          >done</i
+        >
+        <mat-spinner
+          *ngIf="isProcessingVulnerabilityAcceptance"
+          class="ml-2"
+          diameter="20"
+        ></mat-spinner>
+      </button>
+    </div>
+  </div>
 </ng-container>
