+++ admin/webapp/websrc/app/routes/vulnerabilities/vulnerability-items/vulnerability-items-table/vulnerability-items-table.component.html
@@ -9,18 +9,6 @@
     {{ totalCount }}
   </span>
   <div class="d-flex justify-content-end align-items-center">
-    <div
-      *appDisplayControl="'write_vuls_profile'"
-      matTooltip="{{ 'cveProfile.CRD_ACCEPT' | translate }}"
-      [matTooltipDisabled]="cfgType !== 'ground'">
-      <button
-        (click)="acceptVul()"
-        *ngIf="selectedVulnerability"
-        [disabled]="cfgType === 'ground'"
-        mat-stroked-button>
-        <i class="eos-icons mr-2">done</i>{{ 'cveProfile.ACCEPT' | translate }}
-      </button>
-    </div>
     <button
       [matMenuTriggerFor]="menu"
       class="mx-2 d-flex align-items-center justify-content-center<i class='eos-icons'>done</i>a"
