+++ admin/webapp/websrc/app/routes/vulnerabilities/vulnerability-items/vulnerability-items-table/vulnerability-items-table.component.ts
@@ -7,7 +7,6 @@ import {
   Output,
 } from '@angular/core';
 import {
-  CfgType,
   ErrorResponse,
   RemoteGridApi,
   VulnerabilityProfile,
@@ -48,7 +47,6 @@ export class VulnerabilityItemsTableComponent
   @Output() toggleChartView = new EventEmitter();
   @Output() vulnerabilitySelected = new EventEmitter();
   domains!: string[];
-  cfgType!: CfgType;
   selectedVulnerability;
   matchTypes = this.vulnerabilitiesFilterService.matchTypes;
   dateTypes = this.vulnerabilitiesFilterService.dateTypes;
@@ -146,9 +144,9 @@ export class VulnerabilityItemsTableComponent
       },
       cellClass: ['d-flex', 'align-items-center', 'justify-content-end'],
       headerValueGetter: () => this.translate.instant('setting.ACTIONS'),
-      width: 100,
-      maxWidth: 100,
-      minWidth: 100,
+      width: 150,
+      maxWidth: 160,
+      minWidth: 150,
     },
   ];
 
@@ -163,9 +161,6 @@ export class VulnerabilityItemsTableComponent
   ) {}
 
   ngOnInit(): void {
-    this.vulnerabilitiesService.getProfileType().subscribe(cfgType => {
-      this.cfgType = cfgType;
-    });
     this.vulnerabilitiesService.getDomain().subscribe(domains => {
       this.domains = domains;
     });
@@ -218,43 +213,6 @@ export class VulnerabilityItemsTableComponent
           };
         })
       );
-  }
-
-  acceptVul() {
-    const payload: VulnerabilityProfile = {
-      entries: [
-        {
-          name: this.selectedVulnerability.name,
-          days: 0,
-          comment: `Accepted by ${
-            GlobalVariable.user.token.username
-          } at ${this.datePipe.transform(
-            new Date(),
-            'MMM dd, y HH:mm:ss'
-          )} from Vulnerabilities page`,
-          images: [],
-          domains: [],
-        },
-      ],
-      name: 'default',
-    };
-    this.vulnerabilitiesService.acceptVulnerability(payload).subscribe({
-      complete: () => {
-        this.vulnerabilitiesService.refreshing$.next(true);
-        setTimeout(() => {
-          this.vulnerabilitiesService.refresh();
-          this.notificationService.open(
-            this.translate.instant('cveProfile.msg.ADD_OK')
-          );
-        }, 3000);
-      },
-      error: ({ error }: { error: ErrorResponse }) => {
-        this.notificationService.openError(
-          error,
-          this.translate.instant('cveProfile.msg.ADD_NG')
-        );
-      },
-    });
   }
 
   changeScoreView(val: string) {
